\name{teeth}
\alias{teeth}
\title{A simulated data set to demonstrate the zicounts package}
\description{
   'teeth' data is a computer simulated data set to demonstrate the 'zicounts'package
}

\examples{
## See ?zicouts, zicensor 
}

\keyword{models}
\author{S. M. Mwalili <samuelmwalili@yahoo.com>}
